requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-google-genai
langchain-chroma
langchain-community
langchain-core
langgraph
pypdf
httpx
typing_extensions
google-generativeai
tiktoken
pytz
SQLAlchemy
aiosqlite
python-multipart
langchain-ollama
greenlet
rank_bm25
pymupdf
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pydantic-settings
orjson
psutil
aiomysql
cryptography